Move MakeYourDream component to 'Layout'
